Speedup trace loading by properly deduplicationg stacks, fix .net profiling

+++ src/IRExplorerUI/Compilers/FunctionDebugInfo.cs
@@ -109,7 +109,7 @@ public class FunctionDebugInfo : IEquatable<FunctionDebugInfo>, IComparable<Func
     // it is often the slowest part in processing a trace, while the memory
     // saving are quite small (under 15%, a few dozen MBs even for big traces).
     Name = name;
-    RVA = (uint)rva;
+    RVA = rva;
     Size = (uint)size;
     OptimizationLevel = optLevel;
     SourceLines = null;
@@ -126,7 +126,7 @@ public class FunctionDebugInfo : IEquatable<FunctionDebugInfo>, IComparable<Func
   [ProtoMember(4)]
   public long AuxiliaryId { get; set; } // Used for RejitID in managed code.
   [ProtoMember(5)]
-  public uint RVA { get; set; }
+  public long RVA { get; set; }
   [ProtoMember(6)]
   public uint Size { get; set; }
   [ProtoMember(7)]
@@ -140,8 +140,8 @@ public class FunctionDebugInfo : IEquatable<FunctionDebugInfo>, IComparable<Func
   //? TODO: Remove SourceFileName from SourceLineDebugInfo
   public string SourceFileName { get; set; }
   public string OriginalSourceFileName { get; set; }
-  public uint StartRVA => RVA;
-  public uint EndRVA => RVA + Size - 1;
+  public long StartRVA => RVA;
+  public long EndRVA => RVA + Size - 1;
   public bool IsUnknown => RVA == 0 && Size == 0;
 
   public static T BinarySearch<T>(List<T> ranges, long value) where T : IComparable<long> {
